SlimerJS, path normalisation and other bugs
Support for Firefox added. This requires a unique profile to be used on
concurrent processes. SlimerJS does do this by default but it seems only
when there aren't any existing xulrunner processes already running. To
get around this I am creating a temporary profile per slimerjs instance
and then deleting these once the script is finished. On configuration
with a large number of sizes or URL's this folder may grow to a large
size!.

Normalised the paths across platforms to account for / and \ differences
on Windows and Unix

Also included:
- Added ability to specify user agent
- Console output is now displayed
- A few typos and formatting issues resolved